refactor: coordinate 패키지 추가
