+++ src/main/kotlin/minesweeper/domain/cell/coordinate/Row.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain.cell
+package minesweeper.domain.cell.coordinate
 
 @JvmInline
 value class Row(
