+++ src/test/kotlin/minesweeper/domain/MineBoardTest.kt
@@ -8,8 +8,6 @@ import io.kotest.matchers.throwable.shouldHaveMessage
 import minesweeper.domain.MineBoard.Companion.generateNewMineBoard
 import minesweeper.domain.cell.Cell
 import minesweeper.domain.cell.CellType
-import minesweeper.domain.cell.Column
-import minesweeper.domain.cell.Row
 
 class MineBoardTest : FunSpec({
